[cross-platform] fix build with msys2 on windows (2)

+++ src/sysdeps/uio.h
@@ -22,7 +22,7 @@
 
 // Layout of this struct must match struct WSABUF (verified via static assert in sysdeps_win32.cpp)
 struct adb_iovec {
-    size_t iov_len;
+    unsigned long iov_len;
     void* iov_base;
 };
